// src/api/generated/models/apiadmin-stats.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { APIRestUserDB } from './apirest-user-db';

/**
 * 
 * @export
 * @interface APIAdminStats
 */
export interface APIAdminStats {
    /**
     * 
     * @type {number}
     * @memberof APIAdminStats
     */
    'uptime_seconds': number;
    /**
     * 
     * @type {number}
     * @memberof APIAdminStats
     */
    'user_count': number;
    /**
     * 
     * @type {number}
     * @memberof APIAdminStats
     */
    'clan_count': number;
    /**
     * 
     * @type {string}
     * @memberof APIAdminStats
     */
    'last_players_update': string;
    /**
     * 
     * @type {string}
     * @memberof APIAdminStats
     */
    'last_clan_update': string;
    /**
     * 
     * @type {number}
     * @memberof APIAdminStats
     */
    'count_active_users': number;
    /**
     * 
     * @type {Array<APIRestUserDB>}
     * @memberof APIAdminStats
     */
    'active_users_list': Array<APIRestUserDB>;
    /**
     * 
     * @type {number}
     * @memberof APIAdminStats
     */
    'external_api_calls': number;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof APIAdminStats
     */
    'custom_api_calls': { [key: string]: number; };
    /**
     * 
     * @type {Array<{ [key: string]: any; }>}
     * @memberof APIAdminStats
     */
    'last_1000_logs': Array<{ [key: string]: any; }>;
}